pyproject.toml
[tool.poetry]
name = "demoapp"
version = "0.1.0"
description = ""
authors = ["Percy Tejada Donayre <percytejada.code@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langchain = "^0.3.2"
langchain-openai = "^0.2.2"
pydantic = "^2.9.2"
langchain-community = "^0.3.2"
pypdf = "^5.0.1"
langchain-chroma = "^0.1.4"
faiss-cpu = "^1.9.0"
ipykernel = "^6.29.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
